разбивает редьюсер на модули
