+++ project/src/components/pages/film-screen/film-screen.jsx
@@ -168,8 +168,8 @@ FilmScreen.propTypes = {
   authorizationStatus: PropTypes.string.isRequired,
 };
 
-const mapStateToProps = (state) => ({
-  authorizationStatus: state.authorizationStatus,
+const mapStateToProps = ({USER}) => ({
+  authorizationStatus: USER.authorizationStatus,
 });
 
 export { FilmScreen };
