+++ project/src/components/pages/player-screen/player-screen.jsx
@@ -58,8 +58,8 @@ PlayerScreen.propTypes = {
   films: PropTypes.arrayOf(filmProp).isRequired,
 };
 
-const mapStateToProps = (state) => ({
-  films: state.films,
+const mapStateToProps = ({DATA}) => ({
+  films: DATA.films,
 });
