+++ project/src/components/ui/user-block/user-block.jsx
@@ -65,9 +65,9 @@ UserBlock.propTypes = {
   onLogout: PropTypes.func.isRequired,
 };
 
-const mapStateToProps = (state) => ({
-  authorizationStatus: state.authorizationStatus,
-  authInfo: state.authInfo,
+const mapStateToProps = ({USER}) => ({
+  authorizationStatus: USER.authorizationStatus,
+  authInfo: USER.authInfo,
 });
 
 const mapDispatchToProps = (dispatch) => ({
