+++ project/src/store/action.js
@@ -2,9 +2,9 @@ export const ActionType = {
   LOAD_FILMS: 'data/loadFilms',
   LOAD_PROMO_FILM: 'data/loadPromoFilm',
   LOAD_FAVORITE_FILMS: 'data/loadFavoriteFilms',
+  GET_FILMS_LIST: 'data/getFilmsList',
   CHANGE_ACTIVE_GENRE: 'filter/changeActiveGenre',
   RESET_ACTIVE_GENRE: 'filter/resetActiveGenre',
-  GET_FILMS_LIST: 'filter/getFilmsList',
   GET_FILMS_RENDERED_COUNT: 'filter/getFilmsRenderedCount',
   RESET_FILMS_RENDERED_COUNT: 'filter/resetFilmsRenderedCount',
   REQUIRED_AUTHORIZATION: 'user/requiredAuthorization',
